Reference updated with new JSON files after switch to JDK 17
With Processing 4 we switched from JDK 11 to JDK 17. Unfortunately, this caused the script that parses JavaDoc comments in the Processing source and creates JSON files in the website repository to break. As a result, we were unable to update the reference until we had an updated script. However, we have now generated new JSON files using an updated script that can be found at this link: https://github.com/processing/processing-doclet/pull/2